### 2025-01-10 #7645 @editorialbot
[PRE REVIEW]: gigs: A package for standardising fetal, neonatal, and child growth assessment
**Submitting author:** @simpar1471 (<a href="http://orcid.org/0009-0003-8214-4496">Simon Parker</a>)
**Repository:** https://www.github.com/ropensci/gigs
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.1
