### 2025-01-10 #7655 @Kevin-Mattheus-Moerman
Note, this is an ROpenSci related submission (https://github.com/ropensci/software-review/issues/626)
