### 2025-01-10 #7655 @Kevin-Mattheus-Moerman
@simpar1471 Dear author, thanks for this submission. I am the AEiC on this track and will also be assisting with processing the acceptance of this work for JOSS. Before we proceed, please can you have a look at the following points:

- [ ] In your affiliations please spell out all country names so do not use acronyms like UK and USA.
- [ ] Have you created a ZENODO archived version yet? If so please report the DOI here.
- [ ] In the paper check `softwares`, I believe this should be `software` (which can serve as a plural there too).
- [ ] Perhaps remove the comma in `e.g.,`
- [ ] The paper title seems to include the inverted comma's for the PDF.  You currently encode for it like this:
```
title: |
  'gigs: A package for standardising fetal, neonatal, and child growth assessment'
```
Can you try this:
```
title: `gigs: A package for standardising fetal, neonatal, and child growth assessment`
```
