### 2025-01-10 bcfdbfb
chore: update joss paper.md
