### 2025-01-10 #7655 @simpar1471
Thanks for looking over our submission @Kevin-Mattheus-Moerman.

With respect to each point in your previous comment:

1. The publications with missing DOIs do not have DOIs. The first two are WHO publications with ISBNs ([here](https://www.who.int/publications/i/item/924154693X) and [here](https://www.who.int/publications/i/item/9789241547185)) but no DOI; the third is a [GitHub repo](https://github.com/ki-tools/growthstandards/) with no DOI.
2. The project has contributing guidelines on the repo on GitHub - [is this okay](https://github.com/ropensci/gigs?tab=coc-ov-file#readme)?
3. Fixed in new commit to repo.
4. We haven't got a Zenodo DOI.
5. Changed `softwares` to `software` throughout `paper.md`.
6. Removed the comma after `e.g.`.
7. Fixed the backticks.
