### 2025-01-17 5afe54c
docs: update benchmarking article to include new gigs version for Stata
