+++ vignettes/articles/benchmarking.Rmd
@@ -205,7 +205,7 @@ pkg_str <- function(pkgname) paste("R:", pkgname, packageVersion(pkg = pkgname))
 
 ver_zanthro_stata <- "1.0.2"
 ver_gigs_sas <- "0.1.8"
-ver_gigs_stata <- "1.1.1"
+ver_gigs_stata <- "1.1.3"
 if (!exists("key_pkg_label")) {
   # Generate key_pkg_label if doesn't already exist --> otherwise use cached
   # version in benchmarking.rda
@@ -380,7 +380,7 @@ foreach i in 1 10 100 500 1000 5000 10000 25000 50000 75000 100000 {
 ```
 
 ```{r bp_who_gs_stata, echo = FALSE, eval = FALSE}
-bp_who_gs_stata <- read.csv("exclude/statabench/stata_bench.csv")
+bp_who_gs_stata <- read.csv("exclude/statabench/statabench.csv")
 bp_who_gs_stata <- bp_who_gs_stata[bp_who_gs_stata$acronym == "who_gs", ]
 bp_who_gs_stata <- bp_who_gs_stata[, -4]
 ```
@@ -501,7 +501,7 @@ discrepancies_who_gs <- data.frame(z = c(-3.03, -2.97, 2.97, 3.03),
   )
 ```
 
-```{r discrepancies_who_gs_eval, echo = FALSE, eval = TRUE}
+```{r discrepancies_who_gs_eval, echo = FALSE}
 discrepancies_who_gs
 ```
 
@@ -587,7 +587,7 @@ foreach i in 1 10 100 500 1000 5000 10000 25000 50000 75000 100000 {
 ```
 
 ```{r bp_ig_nbs_stata, echo = FALSE, eval = FALSE}
-bp_ig_nbs_stata <- read.csv("exclude/statabench/stata_bench.csv")
+bp_ig_nbs_stata <- read.csv("exclude/statabench/statabench.csv")
 bp_ig_nbs_stata <- bp_ig_nbs_stata[bp_ig_nbs_stata$acronym == "ig_nbs", ]
 bp_ig_nbs_stata <- bp_ig_nbs_stata[, -4]
 ```
@@ -728,7 +728,7 @@ foreach i in 1 10 100 500 1000 5000 10000 25000 50000 75000 100000 {
 ```
 
 ```{r bp_ig_png_stata, echo = FALSE, eval = FALSE}
-bp_ig_png_stata <- read.csv("exclude/statabench/stata_bench.csv")
+bp_ig_png_stata <- read.csv("exclude/statabench/statabench.csv")
 bp_ig_png_stata <- bp_ig_png_stata[bp_ig_png_stata$acronym == "ig_png", ]
 bp_ig_png_stata <- bp_ig_png_stata[, -4]
 ```
@@ -843,7 +843,7 @@ foreach i in 1 10 100 500 1000 5000 10000 25000 50000 75000 100000 {
 ```
 
 ```{r bp_ig_fet_stata, echo = FALSE, eval = FALSE}
-bp_ig_fet_stata <- read.csv("exclude/statabench/stata_bench.csv")
+bp_ig_fet_stata <- read.csv("exclude/statabench/statabench.csv")
 bp_ig_fet_stata <- bp_ig_fet_stata[bp_ig_fet_stata$acronym == "ig_fet", ]
 bp_ig_fet_stata <- bp_ig_fet_stata[, -4]
 ```
@@ -974,57 +974,56 @@ save(file = file.path("vignettes/articles/benchmarking.rda"),
 ## Stata code
 
 ```{stata, eval = FALSE}
-// Set path to directory with benchmarking datasets
-local path = "D:\path\to\directory"
-cd "`path'"
+clear all
 
[…]
