### 2025-01-17 a4af2be
chore: update paper.md for minor readability changes
