### 2025-01-17 #7655 @simpar1471
I just threw a few edits on the pile for readability. As far as I could see, none conflict with your suggested changes.
