### 2025-01-17 0ca699b
docs: update README with JOSS badge
