### 2025-01-28 #7655 @Kevin-Mattheus-Moerman
@simpar1471 thanks for making those changes.

- [ ] I see you now have a code of conduct, but no contributing guidelines yet. Can you look into this?
- [ ] The title is still not fixed. 👆
- [ ] Please create a tagged release for your software and report the version tag here.
- [ ] Next please archive that release on ZENODO and report back here with the DOI.
