### 2025-01-29 #7655 @simpar1471
@Kevin-Mattheus-Moerman

1. There are contributing guidelines at `.github/CONTRIBUTING.md`. Are these sufficient?
2. I've fixed the title - I'd changed to commas but left in the new-line stuff.

Working on doing chores for the new release + archiving on Zenodo.
