### 2025-01-29 d1cfe8e
chore: updates to facilitate new release (0.5.2)
