+++ CITATION.cff
@@ -8,7 +8,7 @@ message: 'To cite package "gigs" in publications use:'
 type: software
 license: GPL-3.0-or-later
 title: 'gigs: Assess Fetal, Newborn, and Child Growth with International Standards'
-version: 0.5.1.9000
+version: 0.5.2
 identifiers:
 - type: url
   value: https://docs.ropensci.org/gigs
@@ -35,7 +35,7 @@ preferred-citation:
   - family-names: Ohuma
     given-names: Eric O
   year: '2024'
-  notes: 0.5.1
+  notes: 0.5.2
   url: https://github.com/ropensci/gigs/
 repository-code: https://github.com/ropensci/gigs
 url: https://github.com/ropensci/gigs/
