+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: gigs
 Title: Assess Fetal, Newborn, and Child Growth with International Standards
-Version: 0.5.1.9000
+Version: 0.5.2
 Authors@R: c(
     person(given = c("Simon", "R"), family = "Parker",
            email = "simon.parker1471@outlook.com",
