### 2025-01-29 #7655 @simpar1471
Zenodo archiving is now live with the latest release - v0.5.2:

[![DOI](https://zenodo.org/badge/627397579.svg)](https://doi.org/10.5281/zenodo.14763685)
