+++ NEWS.md
@@ -1,3 +1,5 @@
+# gigs (development version)
+
 # gigs 0.5.2
 
 - More changes to online articles to facilitate rOpenSci website building.
