+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Type: Package
 Package: gigs
 Title: Assess Fetal, Newborn, and Child Growth with International Standards
-Version: 0.5.2
+Version: 0.5.2.9000
 Authors@R: c(
     person(given = c("Simon", "R"), family = "Parker",
            email = "simon.parker1471@outlook.com",
