### 2025-01-31 f0ab71a
chore: update paper.md to have apostrophes, not backticks, and fix capitalisation of R package names in paper.bib
