+++ paper/paper.bib
@@ -304,7 +304,7 @@ Conclusions This study presents international standards to measure symphysis-fun
 }
 
 @Manual{r_anthro,
-    title = {anthro: Computation of the WHO Child Growth Standards},
+    title = {{anthro}: Computation of the WHO Child Growth Standards},
     author = {Dirk Schumacher},
     year = {2023},
     note = {R package version 1.0.1},
@@ -313,7 +313,7 @@ Conclusions This study presents international standards to measure symphysis-fun
   }
 
 @Manual{r_AGD,
-  title = {AGD: Analysis of Growth Data},
+  title = {{AGD}: Analysis of Growth Data},
   author = {Stef {van Buuren}},
   year = {2024},
   note = {R package version 0.43},
@@ -322,7 +322,7 @@ Conclusions This study presents international standards to measure symphysis-fun
 }
 
 @Manual{r_childsds,
-  title = {childsds: Data and Methods Around Reference Values in Pediatrics},
+  title = {{childsds}: Data and Methods Around Reference Values in Pediatrics},
   author = {Mandy Vogel},
   year = {2022},
   note = {R package version 0.8.0},
@@ -331,7 +331,7 @@ Conclusions This study presents international standards to measure symphysis-fun
 }
 
 @Manual{r_growthstandards,
-  title = {growthstandards: Anthropometric Growth Standard Calculations},
+  title = {{growthstandards}: Anthropometric Growth Standard Calculations},
   author = {Ryan Hafen},
   year = {2024},
   note = {R package version 0.1.6, commit d1fc9ebf1a9008c6b383aba375f43e57b57ad2b7},
@@ -339,7 +339,7 @@ Conclusions This study presents international standards to measure symphysis-fun
 }
 
 @Manual{r_intergrowth,
-  title = {intergrowth: An R Implementation of the INTERGROWTH-21st Standards and Tools},
+  title = {{intergrowth}: An R Implementation of the INTERGROWTH-21st Standards and Tools},
   author = {{Ernest Guevarra}},
   year = {2023},
   note = {R package version 0.0.0.9001},
@@ -348,7 +348,7 @@ Conclusions This study presents international standards to measure symphysis-fun
 }
 
 @Manual{r_sitar,
-  title = {sitar: Super Imposition by Translation and Rotation Growth Curve Analysis},
+  title = {{sitar}: Super Imposition by Translation and Rotation Growth Curve Analysis},
   author = {Tim Cole},
   year = {2023},
   note = {R package version 1.4.0},
@@ -357,7 +357,7 @@ Conclusions This study presents international standards to measure symphysis-fun
 }
 
 @Manual{r_zscorer,
-  title = {zscorer: Child Anthropometry z-Score Calculator},
+  title = {{zscorer}: Child Anthropometry z-Score Calculator},
   author = {Mark Myatt and Ernest Guevarra},
   year = {2019},
   note = {R package version 0.3.1},
