### 2025-01-31 #7655 @simpar1471
@Kevin-Mattheus-Moerman to address the broken backticks, I looked at previously accepted JOSS papers and saw they put their titles in single quotes - so I've updated `paper.md` to have this:
```yaml
title: 'gigs: A package for standardising fetal, neonatal, and child growth assessment'
```
Which compiles with formatting similar to previously published articles.
