### 2025-02-04 71c1710
chore: update docs + CITATION to reflect JOSS publciation
