+++ NEWS.md
@@ -1,5 +1,7 @@
 # gigs (development version)
 
+- Documentation updates due to JOSS publication.
+
 # gigs 0.5.2
 
 - More changes to online articles to facilitate rOpenSci website building.
