+++ CITATION.cff
@@ -9,6 +9,7 @@ type: software
 license: GPL-3.0-or-later
 title: 'gigs: Assess Fetal, Newborn, and Child Growth with International Standards'
 version: 0.5.2.9000
+doi: 10.21105/joss.07655
 identifiers:
 - type: url
   value: https://docs.ropensci.org/gigs
@@ -23,19 +24,30 @@ authors:
   email: simon.parker1471@outlook.com
   orcid: https://orcid.org/0009-0003-8214-4496
 preferred-citation:
-  type: manual
-  title: 'gigs: Assess Fetal, Newborn, and Child Growth with International Standards'
+  type: article
+  title: 'gigs: A package for standardising fetal, neonatal, and child growth assessment'
   authors:
   - family-names: Parker
     given-names: Simon R
-    email: simon.parker1471@outlook.com
     orcid: https://orcid.org/0009-0003-8214-4496
+    email: simon.parker1471@outlook.com
   - family-names: Vesel
     given-names: Linda
+    orcid: https://orcid.org/0000-0003-3753-4172
   - family-names: Ohuma
     given-names: Eric O
-  notes: 0.5.2.9000
-  url: https://github.com/ropensci/gigs/
+    orcid: https://orcid.org/0000-0002-3116-2593
+  journal: Journal of Open Source Software
+  publisher:
+    name: Open Journals
+  year: '2025'
+  month: '2'
+  volume: '10'
+  issue: '106'
+  doi: 10.21105/joss.07655
+  issn: 2475-9066
+  notes: gigs v0.5.2.9000
+  start: '7655'
 repository-code: https://github.com/ropensci/gigs
 url: https://github.com/ropensci/gigs/
 contact:
@@ -52,6 +64,21 @@ keywords:
 - rstats
 - who
 references:
+- type: manual
+  title: 'gigs: Assess Fetal, Newborn, and Child Growth with International Standards'
+  authors:
+  - family-names: Parker
+    given-names: Simon R
+    orcid: https://orcid.org/0009-0003-8214-4496
+  - family-names: Vesel
+    given-names: Linda
+    orcid: https://orcid.org/0000-0003-3753-4172
+  - family-names: Ohuma
+    given-names: Eric O
+    orcid: https://orcid.org/0000-0002-3116-2593
+  notes: gigs v0.5.2.9000
+  doi: 10.5281/zenodo.14763686
+  url: https://github.com/ropensci/gigs/
 - type: software
   title: 'R: A Language and Environment for Statistical Computing'
   notes: Depends
