### 2025-02-04 15fc5f3
chore: fix citation in README.Rmd
