+++ README.Rmd
@@ -485,9 +485,10 @@ Health Centre, London School of Hygiene & Tropical Medicine
 
 ### Citation
 
-Parker SR, Vesel L, Ohuma EO (2025). *gigs: gigs: A package for standardising
-fetal, neonatal, and child growth assessment*. **Journal of Open Source
-Software**, 10(106), 7655. <https://doi.org/10.21105/joss.07655>.
+Parker SR, Vesel L, Ohuma EO (2025). **gigs: A package for standardising
+fetal, neonatal, and child growth assessment**. *Journal of Open Source
+Software*, **10(106):7655**. doi:
+[10.21105/joss.07655](https://doi.org/10.21105/joss.07655).
 
 ## Code of Conduct
 
