### 2025-02-19 35f8cd8
chore: fix citation in README.Rmd
