+++ NEWS.md
@@ -1,6 +1,7 @@
 # gigs (development version)
 
 - Documentation updates due to JOSS publication.
+- Other docs updates to fix minor errors.
 
 # gigs 0.5.2
 
