### 2025-06-27 647e7ef
docs: fix links to vctrs recycling rules (prevent NOTE on rOpenSci CI)
