+++ R/conversion.R
@@ -76,8 +76,8 @@
 #'   [`gigs::ig_png`][ig_png] documentation
 #' * For the WHO Growth standards see the [`gigs::who_gs`][who_gs] documentation
 #' @note Inputs other than `acronym` and `family` will be recycled by
-#'   [vctrs::vec_recycle_common()], and must adhere to the [vctrs] recycling
-#'   rules.
+#'   [vctrs::vec_recycle_common()], and must adhere to the
+#'   \link[=vctrs::vector_recycling_rules]{`vctrs` recycling rules}.
 #' @seealso [report_units()] will tell you the units needed for `y` and `x`, for
 #'   any combination of `family` and `acronym`.
 #' @returns Numeric vector of z-scores/centiles with length equal to the longest
