+++ man/compute_svn.Rd
@@ -62,7 +62,8 @@ appropriate-for-gestational age; LGA, large-for-gestational age.
 }
 \note{
 Input vectors are recycled by \code{\link[vctrs:vec_recycle]{vctrs::vec_recycle_common()}}, and must
-adhere to the \link{vctrs} recycling rules.
+adhere to the
+\link[=vctrs::vector_recycling_rules]{\code{vctrs} recycling rules}.
 }
 \examples{
 compute_svn(
